load a specific version of jQuery
Sometimes you need a specific version of jQuery. This snippet creates a
global variable that contains all versions of jQuery (excluding betas
and a, b) and a global function `loadjQuery(version)` that loads the
specified version or latest if version is not contained in the allowable
versions.

It appends only one script tag to the body no matter how many times it
is called. Most of the time it doesn't manipulate the DOM if the version
you want to load is the version already available.